use global fetch directly

+++ scripts/download-libevdev.mjs
@@ -1,17 +1,18 @@
 import Fs from 'node:fs'
-import { once } from 'node:events'
+import Stream from 'node:stream'
 import C from './util/common.js'
-import { fetch } from './util/fetch.js'
 import * as Tar from 'tar'
 
 const url = `https://github.com/${C.libevdev.owner}/${C.libevdev.repo}/releases/download/v${C.libevdev.version}/${C.libevdev.assetName}`
 
 console.log("fetch", url)
 const response = await fetch(url)
+if (!response.ok) { throw new Error(`bad status code ${response.status}`) }
 
 console.log("unpack to", C.dir.libevdev)
 await Fs.promises.rm(C.dir.libevdev, { recursive: true }).catch(() => {})
 await Fs.promises.mkdir(C.dir.libevdev, { recursive: true })
-const tar = Tar.extract({ gzip: true, C: C.dir.libevdev })
-response.stream().pipe(tar)
-await once(tar, 'finish')
+await Stream.promises.pipeline(
+	Stream.Readable.fromWeb(response.body),
+	Tar.extract({ gzip: true, C: C.dir.libevdev }),
+)
